[BUILD] Propagate `FUSILLI_ENABLE_AMDGPU` compile definitions
CMake's `add_definitions` adds compile definitions to targets in the
current file - which works locally. `target_compile_definitions` adds
the compile definition to all consumers who use the `fusilli` library
CMake target (in the build tree, or when using an installed Fusilli).
We want the latter. Also - in case consumers are cross compiling - we
want to add the compile definition regardless of presence or absence of
AMDGPU on the build machine.